fix: add required socket property to module.json
The socket property must be set to true in the manifest for Foundry VTT
to allow the module to use socket communication. This is a prerequisite
for any module using game.socket.emit() and game.socket.on().

Without this property, socket events cannot be sent or received, which
explains why the socket.on handler was not functioning.

Ref: https://foundryvtt.wiki/en/development/api/sockets